Naive implementation of problem 3
Works well enough to use the Sieve of Eratosthenes for smaller numbers,
but building up enough primes to get the divisors for a number with
twelve digits takes far too long. Switching this to use a probabilistic
method for determining primeness would be more efficient.

Since the prompt is just to find the **largest** prime factor, I wonder if
the best approach is to start at sqrt(n) and work backwards?